+++ src/app/app_priority_lab.c
@@ -29,9 +29,10 @@ void vLowPriorityTask(void *pv)
     vTaskDelay(pdMS_TO_TICKS(3000));  // 開機後等待穩定
     for (;;)
     {
-        printf("[LP][T=%lu] 準備獲取 I2C，預計佔用 CPU 2000ms...\n", xTaskGetTickCount());
+        printf("[LP][T=%lu] 準備獲取 I2C，預計佔用 CPU 2000ms...\n",
+               (unsigned long)xTaskGetTickCount());
         hal_i2c_lab_simulate_long_transfer(2000);
-        printf("[LP][T=%lu] 完成 I2C 操作並釋放鎖。\n", xTaskGetTickCount());
+        printf("[LP][T=%lu] 完成 I2C 操作並釋放鎖。\n", (unsigned long)xTaskGetTickCount());
         vTaskDelay(pdMS_TO_TICKS(10000));  // 實驗結束，無限期休息
     }
 }
@@ -42,9 +43,11 @@ void vMidPriorityTask(void *pv)
     vTaskDelay(pdMS_TO_TICKS(4000));  // 在 LP 執行到一半時殺出
     for (;;)
     {
-        printf("    [MP_%d][T=%lu] 啟動！雙核霸佔開始 (3000ms)...\n", core_id, xTaskGetTickCount());
+        printf("    [MP_%d][T=%lu] 啟動！雙核霸佔開始 (3000ms)...\n", core_id,
+               (unsigned long)xTaskGetTickCount());
         simulate_heavy_work(3000);
-        printf("    [MP_%d][T=%lu] 運算結束，交出 CPU。\n", core_id, xTaskGetTickCount());
+        printf("    [MP_%d][T=%lu] 運算結束，交出 CPU。\n", core_id,
+               (unsigned long)xTaskGetTickCount());
         vTaskDelay(pdMS_TO_TICKS(10000));
     }
 }
@@ -56,7 +59,7 @@ void vHighPriorityTask(void *pv)
     uint8_t dummy = 0x00;
     for (;;)
     {
-        printf("        [HP][T=%lu] 緊急請求 I2C...\n", xTaskGetTickCount());
+        printf("        [HP][T=%lu] 緊急請求 I2C...\n", (unsigned long)xTaskGetTickCount());
         TickType_t t0 = xTaskGetTickCount();
 
         hal_i2c_status_t status = hal_i2c_write_timeout(0x3C, &dummy, 1, 2000);
@@ -64,13 +67,13 @@ void vHighPriorityTask(void *pv)
 
         if (status == HAL_I2C_OK)
         {
-            printf("        [HP][T=%lu] 獲取 I2C 成功！實際阻塞耗時: %lu ms\n", t1,
-                   (t1 - t0) * portTICK_PERIOD_MS);
+            printf("        [HP][T=%lu] 獲取 I2C 成功！實際阻塞耗時: %lu ms\n", (unsigned long)t1,
+                   (unsigned long)((t1 - t0) * portTICK_PERIOD_MS));
         }
         else
         {
-            printf("        [HP][T=%lu] 獲取 I2C 失敗！阻塞耗時: %lu ms\n", t1,
-                   (t1 - t0) * portTICK_PERIOD_MS);
+            printf("        [HP][T=%lu] 獲取 I2C 失敗！阻塞耗時: %lu ms\n", (unsigned long)t1,
+                   (unsigned long)((t1 - t0) * portTICK_PERIOD_MS));
         }
 
         vTaskDelay(pdMS_TO_TICKS(10000));
